+++ .babelrc
@@ -1,3 +1,15 @@
 {
-  "presets": ["env", "react"]
+  "presets": [
+    [
+      "@babel/preset-env",
+      {
+        "debug": false,
+        "targets": {
+          "browsers": [
+            "last 3 versions"
+          ]
+        }
+      }
+    ]
+  ]
 }
