+++ package.json
@@ -8,27 +8,30 @@
   "license": "MIT",
   "scripts": {
     "lint": "eslint src/**",
-    "test": "jest --testPathPattern ./test/unit/"
+    "test": "jest test"
   },
   "dependencies": {
+    "@babel/polyfill": "^7.0.0",
     "aws-sdk": "^2.325.0",
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-react": "^6.24.1",
     "joi": "^13.7.0",
     "mongoose": "^5.3.0",
     "uuid": "^3.3.2"
   },
   "devDependencies": {
+    "@babel/cli": "^7.1.2",
     "@babel/core": "^7.1.2",
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
+    "@babel/plugin-transform-arrow-functions": "^7.0.0",
+    "@babel/preset-env": "^7.1.0",
+    "babel-core": "^7.0.0-bridge.0",
     "babel-jest": "^23.6.0",
-    "babel-loader": "^8.0.4",
-    "babel-polyfill": "^6.26.0",
     "eslint": "^5.6.1",
     "eslint-config-airbnb-base": "^13.1.0",
     "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-jest": "^21.24.1",
     "jest": "^23.6.0",
-    "node": "^10.11.0",
+    "regenerator-runtime": "^0.12.1",
     "webpack": "^4.20.2",
     "webpack-cli": "^3.1.2",
     "webpack-node-externals": "^1.7.2"
